requirements.txt
boto3>=1.26.0
botocore>=1.29.0
Pillow>=10.0.0
opencv-python>=4.10.0
python-dotenv>=1.0.0
tabulate>=0.9.0
numpy>=2.0.0
